I would be shocked if I actually remembered to document all of the changes that I made here.  Some of them were shortsighted and reversed to something like what CyberianTiger had done before I got to them.
On the whole I think that the api is easier to use now.  For one thing, writing
to the SocketChannel can now be done by wrapping a BufferedWriter around it.
Initially I had tried to setup things so that a BufferedReader could be used
for reading from the channel, ignoring CT's warnings about why he didn't do
that.  Well, he was right there, I was wrong. (But I was right about writing!)

The examples work.  I think I got all of the debugging messages removed.
There are still some things left to be done, primarily documentation, of course,
and probably fixing what types of exceptions are thrown.  CT's right about
it being wrong for IOException to be thrown all the time. I would also suggest
refactoring the packaging towards org.osjava.nio, but that's not that critical.
I would also wait until we move to subversion for source control.  There are
also still some warnings to clean up.

Some of the major things that were done.
 - Setting up the channels is easier now, requiring less code from the end
   developer.
 - To write to a channel wrap the SocketChannelWriter in a BufferedWriter or
   similar.  get the writer via SocketChannelHandler.getChannelWriter()
 - Multiple objects can listen to a channel (and well...write too i suppose)
   They must implemnet SocketListener (not sure I'm happy with that name) and
   register via SocketChannelHandler.addListener().
   These objects must have the receiveData(ByteBuffer) method (enforced by
   the interface.

I'm going to be using this in my other project as it is now, so it may evolve
a bit more as I find errors, but so long as there are not huge objections,
I would like to see this branch merged back into the mainstream.  I won't do it
without input from others (that means you especially CT).  I know the api
got warped a bit, but I would hope that it's for the better.  The idea behind
it was useability.

Ok, now that my rant is over....here is a more detailed list of the changes.
I do not believe that this list is complete.

- ChannelHandler
  - Refactored read() to readFromChannel()
  - Refactored write() to writeToChannel()

- SocketChannelHandler
  - Removed 'Output out' field.
  - Changed the constructor to be SocketChannelHandler(SocketChannel,IOThread)
    - This was done because it was necessary for deregistering the channel.
  - Refactored the old read() method to readFromChannel() and reduced the
    amount of stuff that it does.  All it does now is fill the readbuffer
    from the channel.
  - Refactored the old write() to writeToChannel()
  - writeTo() method removed.
  - removed resizeBuffer().
  - Added addListener/removeListener for listeners to the socket.

- IOThread
  - Moved all register() and deregister() functionality here.

- IOUtils
  - Refactored SocketChannelHandler connect(InetSocketAddress,IOThread,Socket)
    to SocketChannelHandler connect(InetSocketAddress,IOThread)
    - Socket stuffs don't apply anymore.

- SocketChannelHandlerAcceptor
  - acceptSocketChannelHandler now throws an IOException

- AbstractSocketChannelHandlerAcceptor
  - removed registration functionality from acceptSocketChannelHandler()
  - removed setupSocketStream() method.  It no longer applies.
  - removed createStream()

- Removed unused imports across the project.

- Removed Output

- Removed Input

- Removed Stream

- Removed IOStream

- Denoted who made the comment on some of the more colourful ones, and on
  TODO's where I remembered.

- Added SocketChannelWriter
  - Separate class for handling the underlying channel writes, but that
    can also be wrapped in a BufferedWriter.

- Added SocketListener
  - Designed to alert objects listening to the SocketChannel that there is
    data waiting.