Merged the two maps which hold changes to the interestOps for a key into one map which holds the ops which it is to be changed to.
Added extra checks when altering a key's interestOps to ensure it's a
valid key (This currently killed the thread when a connection is dropped
by the remote end).

Removed a line which ensured that once interest in the read operation was
set there was no way to unset it. (The line did nothing as if you're
already interested in read ops there's no point making sure you're interested
in more read operations).