Wrapped the op change used in the SendingByteBroker class in a 'task' to avoid SecurityExceptions for the wrong Thread executing the task of changing ops.
This seems to work reasonably well, though I know that this is not the fix that
CT is going to be all that satisfied with.  It does avoid, however, exposing 
changing of InterestOps in Channels, while still allowing other threads to 
send messages down the channel.  There has been some discussion of implementing
a wrapper method that would take care of this in order to make it easier for 
people to actually use.  When and if that happens, this change may be reverted.