Two changes:
1) Tasks is now a LinkedList, addTask appends to the end, the Iterator
has been replaced with while(tasks.size() > 0) { Runnable r = (Runnable) tasks.remove(0); ..... } This avoids a ConcurrentModificationException in the case where executing a Task causes another Task to be added.

2) When the Runnable is executed, we synchronized on it, then called .notifyAll() on it. This lets the caller do the following:

Runnable r = new MyTask();
synchronized(r) {
    myIOThread.addTask(r);
    wait();
}

The users thread is paused until the task is executed. (Useful in the case where the user needs to wait for the task to be run before continuing.)
There is one minor gotcha with this:

If Thread.currentThread() == myIOThread this will cause the program to hang.