Implemented a basic queue system. Very basic. It uses ultra-fast LinkedList technology.
I switched constructors to be given MemoryXxx's and not Xxx's so that
ClassCastExceptions will be thrown when people try to mix JMS systems.

No Threading for MessageListeners, though it's meant to be asynchronous
so we might find we have to add it. Though we're not looking to pass the
TCK or anything.

Switched a bunch of methods to be defined in terms of similar methods - 
receive/send etc.