A little more rearrangement.
Filled out some of the more basic get and update methods.  This is not nearly
as big of an accomplishment as it looks.  Those methods that look for the 
column based upon the name, now call the method that looks for the column based
upon its index.  The methods that they call are not likely implemented yet.