Added code to allow differences between multiple source and dest jar files. Added code to allow jar files to be specified as a url. Added code to allow the name of the old version and the name of the new version to be set explicitely (with no current defaults). Added code to allow a dependency path to be specified (not currently used, however will be needed for checking for class compatibility.
The one major thing that's still missing is the ability to pass most of these
arguments in as command line arguments. (Also specifying an output type,
which will be little more than an inbuilt xslt stylesheet, probably just
need two stylesheets for this, one which outputs html, and the other which
outputs plain text.)

Also, ant, eclipse, maven integration are still glaring ommissions.