First attempt at an ant task.
It works perfectly however I need to do a lot of work to reduce the
dependencies.

Currently to use the ant task, you need commons-cli. 
This should not be the case.

Also, it calls javax.xml.* and javax.transformer.* directly which means
that ant will not accept it as a standard task (ant tasks need to be
compatible to java 1.2 according to the website).

Basically what's left that needs doing:
Move JarDiff.main() to another class to remove dep on commons-cli.
Remove javax.xml and javax.transformer deps from DOMDiffHandler.
Make JarDiffTask use Ant's classes to get a handle on the dom builder and
transformers.