Moved main method to Main.java (removes dependency on commons-cli for ant-task)
Added StreamDiffHandler.java, this is a diff handler which writes out an xml
file, which does not need any xml libraries. This is used by the ant task.
(This is for backward compatability, and to avoid the nightmare that is 
javax.xml.* and javax.transform.* from within an ant task).

Deleted SAXDiffHandler, no longer used, and I have no inclination to maintain
two seperate DiffHandlers when I change the API.

Added methods to send a list of contents to DiffHandler.
Implemented in DOMDiffHandler and StreamDiffHandler.
Changed JarDiff to use the new API.

Removed the formating options from JarDiffTask. (Users are expected to 
use the ant built in XSLT task for this, I will provide a subclassed 
XSLTProcess which allows the stylesheet specified to be a url or a resource.

At some point I'll probably approach the ant peoples for incorporating these
changes into their xslt task. (Using a url to a stylesheet is always useful,
using a resource in the classpath is parhaps less so, but still).

Changed project.properties to point at the new Main class instead of at JarDiff.

Altered the text and html stylesheets to filter out the contents of the old and
new jar files, as this is not useful (other than for linking to api docs, which
I have not yet implemented).