Err, not sure about total changes.
xml output now uses internal java names. (this is needed for a number of reasons).
Huge modifications to the html stylesheet (now much much more readable).

Added arguments to specify from api and to api locations, also added ability to specify a custom stylesheet. (This is trivial from the ant task as the built in xslt task with my styleurl hack already supports this).